Renamed react-filepond.d.ts to index.d.ts to allow dtslint-ing
